Rename remote folder to remotes
