+++ webview/src/setup/components/App.tsx
@@ -9,7 +9,7 @@ import { Dvc } from './dvc/Dvc'
 import { Experiments } from './experiments/Experiments'
 import { Studio } from './studio/Studio'
 import { SetupContainer } from './SetupContainer'
-import { Remotes } from './remote/Remotes'
+import { Remotes } from './remotes/Remotes'
 import { useVsCodeMessaging } from '../../shared/hooks/useVsCodeMessaging'
 import { sendMessage } from '../../shared/vscode'
 import { TooltipIconType } from '../../shared/components/sectionContainer/InfoTooltip'
