+++ webview/src/setup/components/remotes/styles.module.scss
